Atualização do controller cozinha adicionando o endpoint do delete

+++ src/main/java/com/algaworks/algafood/api/controller/CozinhaController.java
@@ -4,6 +4,7 @@ import com.algaworks.algafood.domain.repository.CozinhaRepository;
 import com.algaworks.algafood.model.Cozinha;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -59,13 +60,20 @@ public class CozinhaController {
 
     @DeleteMapping("/{cozinhaId}")
     public ResponseEntity<Cozinha> remover(@PathVariable Long cozinhaId) {
-        Cozinha cozinha = cozinhaRepository.porId(cozinhaId);
 
-        cozinhaRepository.remover(cozinha);
+        try {
+            Cozinha cozinha = cozinhaRepository.porId(cozinhaId);
 
-        return ResponseEntity.noContent().build();
+            if (cozinha != null) {
+                cozinhaRepository.remover(cozinha);
 
+                return ResponseEntity.noContent().build();
+            }
 
+            return ResponseEntity.notFound().build();
+        } catch (DataIntegrityViolationException e){
+            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+        }
     }
 
 }
